+++ docs/changelog.rst
@@ -4,6 +4,7 @@ Changelog
 
 Changes in 0.6.X
 ================
+- Fixed Django 1.4 sessions first save data loss
 - FileField now automatically delete files on .delete()
 - Fix for GenericReference to_mongo method
 - Fixed connection regression